[refactor] ProfileHeader ProfileContentView 파일 안에서 생성하는 것으로 수정
